refactor: update import paths for shared utilities in search steps
